requirements.txt
Flask==2.0.1
boto3==1.18.25
pdf2image==1.16.0
pillow==8.3.1
paddleocr==2.3.0.2